requirements.txt
# CQL RLVR — Dependencies
# Core
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0

# Data processing
# (standard library: json, csv, re, hashlib — no extra deps needed)

# NeMo Gym + NeMo RL (install separately on RunPod)
# pip install nemo-gym
# pip install nemo-rl
# Or clone from:
#   git clone https://github.com/NVIDIA-NeMo/Gym.git
#   git clone https://github.com/NVIDIA-NeMo/RL.git
